feat: generalize header

+++ src/interfaces/base/ISigVerifier.sol
@@ -7,7 +7,7 @@ interface ISigVerifier {
         uint48 epoch,
         bytes memory message,
         uint8 keyTag,
-        uint208 quorumThreshold,
+        uint256 quorumThreshold,
         bytes calldata proof
     ) external view returns (bool);
 }
